displace modal logic to Modal component
