+++ packages/example/src/components/modals/Modal.tsx
@@ -1,0 +1,54 @@
+import styled from '@emotion/native';
+import React from 'react';
+import { View, ModalProps } from 'react-native';
+import { Typography } from '../../design-system/components/Typography';
+import { Spacer } from '../../design-system/components/Spacer';
+import { SpatialNavigationRoot } from '../../../../lib/src/spatial-navigation/components/Root';
+import { useLockModal } from '../../hooks/useLockModal';
+
+type CustomModalProps = ModalProps & {
+  isModalVisible: boolean;
+  setIsModalVisible: (isVisible: boolean) => void;
+  children: React.ReactNode;
+  title: string;
+};
+
+export const Modal = ({ isModalVisible, setIsModalVisible, children, title }: CustomModalProps) => {
+  useLockModal({ isModalVisible, setIsModalVisible });
+
+  if (!isModalVisible) return null;
+
+  return (
+    <SpatialNavigationRoot>
+      <StyledModal>
+        <ModalContentContainer>
+          <Typography variant="title" fontWeight="strong">
+            {title}
+          </Typography>
+          <Spacer gap="$8" />
+          {children}
+        </ModalContentContainer>
+      </StyledModal>
+    </SpatialNavigationRoot>
+  );
+};
+
+const StyledModal = styled(View)({
+  position: 'absolute',
+  top: 0,
+  left: 0,
+  right: 0,
+  bottom: 0,
+  justifyContent: 'center',
+  alignItems: 'center',
+});
+
+const ModalContentContainer = styled(View)({
+  minHeight: 200,
+  minWidth: 200,
+  backgroundColor: 'grey',
+  padding: 32,
+  margin: 16,
+  borderRadius: 16,
+  justifyContent: 'center',
+});
