+++ packages/example/src/components/modals/SubtitlesModal.tsx
@@ -1,8 +1,7 @@
-import { SpatialNavigationRoot } from '../../../../lib/src/spatial-navigation/components/Root';
 import { DefaultFocus } from '../../../../lib/src/spatial-navigation/context/DefaultFocusContext';
 import { Button } from '../../design-system/components/Button';
 import { Spacer } from '../../design-system/components/Spacer';
-import { GenericModal } from './GenericModal';
+import { Modal } from './Modal';
 import { useLockModal } from '../../hooks/useLockModal';
 
 interface SubtitlesModalProps {
@@ -19,42 +18,44 @@ export const SubtitlesModal = ({
   useLockModal({ isModalVisible, setIsModalVisible });
 
   return (
-    <SpatialNavigationRoot>
-      <GenericModal isVisible={isModalVisible} title={'Choose subtitles'}>
-        <DefaultFocus>
-          <Button
-            label="English"
-            onSelect={() => {
-              setSubtitles('English');
-              setIsModalVisible(false);
-            }}
-          />
-        </DefaultFocus>
-        <Spacer gap="$8" />
+    <Modal
+      isModalVisible={isModalVisible}
+      setIsModalVisible={setIsModalVisible}
+      title={'Choose subtitles'}
+    >
+      <DefaultFocus>
         <Button
-          label="Spanish"
+          label="English"
           onSelect={() => {
-            setSubtitles('Spanish');
+            setSubtitles('English');
             setIsModalVisible(false);
           }}
         />
-        <Spacer gap="$8" />
-        <Button
-          label="Portuguese"
-          onSelect={() => {
-            setSubtitles('Portuguese');
-            setIsModalVisible(false);
-          }}
-        />
-        <Spacer gap="$8" />
-        <Button
-          label="None"
-          onSelect={() => {
-            setSubtitles('No');
-            setIsModalVisible(false);
-          }}
-        />
-      </GenericModal>
-    </SpatialNavigationRoot>
+      </DefaultFocus>
+      <Spacer gap="$8" />
+      <Button
+        label="Spanish"
+        onSelect={() => {
+          setSubtitles('Spanish');
+          setIsModalVisible(false);
+        }}
+      />
+      <Spacer gap="$8" />
+      <Button
+        label="Portuguese"
+        onSelect={() => {
+          setSubtitles('Portuguese');
+          setIsModalVisible(false);
+        }}
+      />
+      <Spacer gap="$8" />
+      <Button
+        label="None"
+        onSelect={() => {
+          setSubtitles('No');
+          setIsModalVisible(false);
+        }}
+      />
+    </Modal>
   );
 };
